requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
distlib==0.3.9
fastapi==0.115.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
virtualenv==20.27.1
